+++ LICENCE
@@ -1,0 +1,1 @@
+ALL RIGHTS RESERVED TO SHIVANSH PRATAP.
