+++ README.md
@@ -1,3 +1,1 @@
-To START XENETOR  voice code generator Open "Run.bat" file in seprj folder
-To stop  XENETOR Say "stop" (without adding any parse phrase in start or end of stop)  
-Project Under Construction.
+# My First Desktop Applictaion
