+++ src/index.css
@@ -1,0 +1,25 @@
+body {
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica,
+    Arial, sans-serif;
+  margin: auto;
+  /* max-width: 38rem; */
+  padding: 2rem;
+  background-color: rgb(255, 255, 255);
+  color: rgb(0, 0, 0);
+}
+.click {
+  width: 6rem;
+  height: 2rem;
+  background-color: rgb(0, 0, 0);
+  color: rgb(255, 255, 255);
+}
+
+.clicked {
+  background-color: rgb(0, 0, 0);
+  color: rgb(255, 255, 255);
+}
+
+.clicked-btn {
+  background-color: rgb(255, 255, 255);
+  color: rgb(0, 0, 0);
+}
