--- v_input.txt
@@ -1,1 +1,0 @@
-run loop from variable a value 10 to 400 increments where a less then b
